refactor(activator): move result types into steplibrary, drop shim
The activation result types lived in activator/result with a
backwards-compat alias shim in the activator package. Move them to the
steplibrary package (their canonical owner) and update callers to use
steplibrary.ActivatedStep / steplibrary.ActivationType* directly,
removing the shim and the activator/result package.
