+++ activator/git_ref.go
@@ -11,6 +11,7 @@ import (
 	"github.com/bitrise-io/go-utils/command/git"
 	"github.com/bitrise-io/stepman/models"
 	"github.com/bitrise-io/stepman/stepid"
+	"github.com/bitrise-io/stepman/steplibrary"
 	"github.com/bitrise-io/stepman/stepman"
 )
 
@@ -19,10 +20,10 @@ func ActivateGitRefStep(
 	id stepid.CanonicalID,
 	activatedStepDir string,
 	workDir string,
-) (ActivatedStep, error) {
+) (steplibrary.ActivatedStep, error) {
 	repo, err := git.New(activatedStepDir)
 	if err != nil {
-		return ActivatedStep{}, err
+		return steplibrary.ActivatedStep{}, err
 	}
 
 	var cloneCmd *command.Model
@@ -40,21 +41,21 @@ even if the repository is open source!`)
 		}
 		var exitErr *exec.ExitError
 		if errors.As(err, &exitErr) {
-			return ActivatedStep{}, fmt.Errorf("command failed with exit status %d (%s): %w", exitErr.ExitCode(), cloneCmd.PrintableCommandArgs(), errors.New(out))
+			return steplibrary.ActivatedStep{}, fmt.Errorf("command failed with exit status %d (%s): %w", exitErr.ExitCode(), cloneCmd.PrintableCommandArgs(), errors.New(out))
 		}
-		return ActivatedStep{}, err
+		return steplibrary.ActivatedStep{}, err
 	}
 
 	stepYMLPath := filepath.Join(workDir, "current_step.yml")
 	if err := command.CopyFile(filepath.Join(activatedStepDir, "step.yml"), stepYMLPath); err != nil {
-		return ActivatedStep{}, err
+		return steplibrary.ActivatedStep{}, err
 	}
 
-	return ActivatedStep{
+	return steplibrary.ActivatedStep{
 		StepInfo:         models.StepInfoModel{},
 		StepYMLPath:      stepYMLPath,
 		DidStepLibUpdate: false,
-		ActivationType:   ActivationTypeGitRef,
+		ActivationType:   steplibrary.ActivationTypeGitRef,
 		ExecutablePath:   "",
 	}, nil
 }
