+++ activator/path_ref.go
@@ -8,6 +8,7 @@ import (
 	"github.com/bitrise-io/go-utils/pathutil"
 	"github.com/bitrise-io/stepman/models"
 	"github.com/bitrise-io/stepman/stepid"
+	"github.com/bitrise-io/stepman/steplibrary"
 	"github.com/bitrise-io/stepman/stepman"
 )
 
@@ -16,19 +17,19 @@ func ActivatePathRefStep(
 	id stepid.CanonicalID,
 	activatedStepDir string,
 	workDir string,
-) (ActivatedStep, error) {
+) (steplibrary.ActivatedStep, error) {
 	log.Debugf("Local step found: (path:%s)", id.IDorURI)
 	// id.IDorURI is a path to the step dir in this case
 	stepAbsLocalPth, err := pathutil.AbsPath(id.IDorURI)
 	if err != nil {
-		return ActivatedStep{}, err
+		return steplibrary.ActivatedStep{}, err
 	}
 
 	exist, err := pathutil.IsDirExists(stepAbsLocalPth)
 	if err != nil {
-		return ActivatedStep{}, fmt.Errorf("check if a directory exists at %s: %w", stepAbsLocalPth, err)
+		return steplibrary.ActivatedStep{}, fmt.Errorf("check if a directory exists at %s: %w", stepAbsLocalPth, err)
 	} else if !exist {
-		return ActivatedStep{}, fmt.Errorf("the provided directory doesn't exist: %s", stepAbsLocalPth)
+		return steplibrary.ActivatedStep{}, fmt.Errorf("the provided directory doesn't exist: %s", stepAbsLocalPth)
 	}
 
 	log.Debugf("stepAbsLocalPth: %s, stepDir:%s", stepAbsLocalPth, activatedStepDir)
@@ -36,25 +37,25 @@ func ActivatePathRefStep(
 	origStepYMLPth := filepath.Join(stepAbsLocalPth, "step.yml")
 	exist, err = pathutil.IsPathExists(origStepYMLPth)
 	if err != nil {
-		return ActivatedStep{}, fmt.Errorf("check if step.yml exists at %s: %w", origStepYMLPth, err)
+		return steplibrary.ActivatedStep{}, fmt.Errorf("check if step.yml exists at %s: %w", origStepYMLPth, err)
 	} else if !exist {
-		return ActivatedStep{}, fmt.Errorf("step.yml doesn't exist at %s", origStepYMLPth)
+		return steplibrary.ActivatedStep{}, fmt.Errorf("step.yml doesn't exist at %s", origStepYMLPth)
 	}
 
 	activatedStepYMLPath := filepath.Join(workDir, "current_step.yml")
 	if err := command.CopyFile(origStepYMLPth, activatedStepYMLPath); err != nil {
-		return ActivatedStep{}, err
+		return steplibrary.ActivatedStep{}, err
 	}
 
 	if err := command.CopyDir(stepAbsLocalPth, activatedStepDir, true); err != nil {
-		return ActivatedStep{}, err
+		return steplibrary.ActivatedStep{}, err
 	}
 
-	return ActivatedStep{
+	return steplibrary.ActivatedStep{
 		StepInfo:         models.StepInfoModel{},
 		StepYMLPath:      activatedStepYMLPath,
 		DidStepLibUpdate: false,
-		ActivationType:   ActivationTypePathRef,
+		ActivationType:   steplibrary.ActivationTypePathRef,
 		ExecutablePath:   "",
 	}, nil
 }
