+++ activator/steplib_ref_test.go
@@ -8,6 +8,7 @@ import (
 	"github.com/bitrise-io/go-utils/pathutil"
 	"github.com/bitrise-io/stepman/models"
 	"github.com/bitrise-io/stepman/stepid"
+	"github.com/bitrise-io/stepman/steplibrary"
 	"github.com/stretchr/testify/require"
 )
 
@@ -205,9 +206,9 @@ func BenchmarkActivateSteplibRefStep(b *testing.B) {
 					b.Fatal("ActivateSteplibRefStep() succeeded unexpectedly")
 				}
 
-				want := ActivatedStep{
+				want := steplibrary.ActivatedStep{
 					StepYMLPath:      tmpDir + "/current_step.yml",
-					ActivationType:   ActivationTypeSteplibSource,
+					ActivationType:   steplibrary.ActivationTypeSteplibSource,
 					DidStepLibUpdate: !tt.didStepLibUpdateInWorkflow,
 				}
 				require.Equal(b, want, got)
